Minor refactorings, etc.
(test) application.conf: Fixed Issue 19
moved some of the unit tests to be part of it.
