+++ src/it/scala/com/phasmid/majabigwaduce/examples/matrix/MatrixOperation.scala
@@ -8,7 +8,7 @@ import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import akka.util.Timeout
 import com.phasmid.majabigwaduce._
-import com.phasmid.majabigwaduce.examples.CountWords.getTimeout
+import com.phasmid.majabigwaduce.examples.countwords.CountWords.getTimeout
 import com.typesafe.config.{Config, ConfigFactory}
 
 import scala.concurrent.duration._
