+++ src/it/scala/com/phasmid/majabigwaduce/examples/webcrawler/WebCrawlerSpec.scala
