+++ src/it/scala/com/phasmid/majabigwaduce/examples/webcrawler/package.scala
