+++ src/test/resources/application.conf
@@ -35,6 +35,12 @@ akka {
   loggers = ["akka.event.slf4j.Slf4jLogger"]
   logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
 
+  cluster {
+    jmx {
+      multi-mbeans-in-same-jvm = on
+    }
+  }
+
   actor {
     debug {
       # enable function of LoggingReceive, which is to log any received message at
